[SAFE-72] Move casl ability definition to security folder
